requirements.txt
pandas
openpyxl
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib